6645 FIX Solaris Agent: Correctly separate cputime and elapsed time
The process information delivered by the agent has been upgrated multiple
times by requirements of users or information available from the OS. The ps
agent tracked the memory outilization of a process by its name, but could
not access information of process id and the time of life(elapsed time) of
a running process, as such information was misplaced by the agent and thus
ignored by the check. This fix brings this data into consideration by the
check.

Change-Id: Ia6f2df3ee5c074c258ed9a301716c923a3a11451