6860 FIX Fixed random backup issues: "Connection reset by peer"
The "mkbackup" and "omd backup" commands could fail while backing up a site when
the rrdcached process was reloaded, restarted or stopped during the backup. This
could also be triggered by an "Activate changes" action in the GUI.

The backup was then aborted with an error like this: "Failed to perform backup: [Errno 104] Connection reset by peer".

This issue did not affect backups performed for not running sites.

FEED-2982
FEED-3045
CMK-943

Change-Id: I020700f4a9d347a14a79f5440ac2024d8b45e042