6902 FIX apache_status: Ignore certificate for localhost
If a https server at 127.0.0.1 or [::1] is checked, ignore the certificate in
case the name does not match.

Previously we tried to contact the server via http on port 80 instead,
but the server may not be listening on that port.

You need to change the protocol to http if you are monitoring a https
server that also listenes on port 80 and the servers address is not
either the address the certificate was issued for or one of
127.0.0.1, [::1] and "localhost".

Change-Id: I2db8af14aeb238494e558358c07d27f7733fa8cb