Dynamic configuration: Add distributed WATO support
In distributed WATO setups there is only one site where changes to the
Check_MK configuration can be made. This is the central site. However,
there are a lot of situations where the source for the dynamic
configuration, e.g. a Kubernetes instance, is only reachable from a remote
Check_MK site.

This means that we need to split up the DCD connection logic into two
phases.

In phase 1 the connection on the remote site talks to the source system,
e.g. the Kubernetes instance, and creates some kind of "configuration
snapshot".  This information may be a collection of Pods, containers and
so on which will be used to update the Check_MK configuration in the
next step.

The DCD on the central site is communicating with the DCD on the remote
site to get the latest "configuration snapshot". With this knowledge the
phase 2 is launched. In this phase the connection is comparing the phase
1 information with data fetched from the Web API of the central WATO
site and deciding to create, update or delete objects in the Check_MK
configuration. These updates are made immediately to the configuration
and activated once it's allowed.

* The DCD now knows whether or not it's running in a distributed WATO
  remote site
* Split connection execute() into two phases
* Moved phase execution logic to the Connector() base class
* Improved test coverage

CMK-1140

Change-Id: I51e7f0dc1e9df7e9c8c25a0598c3cf94292666ed