Extend docs and validation for regex groups in ps check config
Group matches are instantiated during discovery, to recognize different
services. But at checktime, because the way regex is intended to work
matches are substrings. In the case the user has processes with common
substrings, this would lead to repeated counts of process for the process
rule with the shortest string. This could be the intended behavior. In case
the user wants to have each service considered as separate case at
checktime, he should be informed about using delimiters around groups.

CMK-1276

Change-Id: Ic85218fe70d3696dacc5e84226eac1a8aab9e4e9