6638 FIX if_fortigate: Fix incorrect usage of Name as Alias
Werk #4539 introduced special handling for interfaces of
Fortigate firewalls, but incorrectly substituted the ifName
for the ifAlias, leading to incompatibilities for users
using the alias as an item. Rediscovery is required
to get the proper alias back as an item. If you do not
use the alias, this does not affect you.

Change-Id: I85d394291ce800fe012d91fe59b8e0841674a6ad