6895 FIX hp_msa_volume: Use volume-name as item identifier
Previously the durable-id has been used as identifier of items for the
check hp_msa_volume (and its sub checks io and df). The durable-id is not
guaranteed to be unique for different controllers of a volume. This werk
changes the unique identifier for items from durable-id to volume-name.
As soon as a discovery is executed the hp_msa_volume check items
appear volume-name.

CMK-1129

Change-Id: I72c5630d6c3bdbbadd85bb0a9ba920da69f60c01