6418 FIX Fix a crash during notification cancelling
When a notification was cancelled and the cancelling rule contained
more contacts than the notification that was going to be cancelled
the processing of notifications raised an exception. As a result
no further notifications were processed.

The exception was raised e.g. if a notification for the contact "ab"
was cancelled for the users "ab" and "cd" using the option "The
following users" in the contact selection.

With this werk the issue is fixed.

Change-Id: I725452b02e5db22977707044d0474a4ab4340725