6913 FIX Allow deletion of hosts that are configured to be their own parents
The GUI allowed to configure a host as it's own child and informs the user
about this with a configuration warning message. But it allowed the user to
save the host.

To fix this the user could not delete the host because the GUI prevents deletion
of hosts which are still parent hosts. This check is now skipping the relation
to itself which makes it possible to delete the host without editing the parent
setting.

Change-Id: I4c782429168a514543efa9ad1c930abda0902cf7