6941 FIX mk_logwatch.aix: Don't clear systems error log
The systems error log (errpt) is no longer cleared ("errclear 0").
Previously the errorlog has been cleared after each run of the plugin,
and a backup file of the error log has been created as
/var/log/errpt_$DATE.log.

This is no longer the case, as a less intrusive way to avoid
redundant log messages has been established.

Change-Id: I239d67767e7d5f258fee4616bbd45cd3a89c803f