Fixed dictionary element evaluation during initialization
The Dictionary() valuespec elements can either be a list of elements
or a function reference which is meant to generate the elements on
demand in the moment they are needed instead of the construction time
of the valuespec.

There are attributes optional_keys and required_keys which tell the
Dictionary() valuespec how to handle the single elements.

There are multiple ways to tell the object that none of the elements
is optional. It can be done by either setting it to None, False or
an empty list.

And here comes the bug: When the "optional_keys" attribute was set to an
empty list, the element generation function was called while it wasn't
needed to call it.

In many situations this is not a real problem because mostly the element
function is cheap. But we had a situation where the element function
initialized livestatus connections to all sites during the program
initialization phase where such things should not be done.

Change-Id: I33b4a80aa9ebcf587702f8e1633cbc833c127503