6915 FIX IPMI Management board: Fixed random "Session no longer connected" message
The Check_MK service of hosts that have a IPMI based management board configured could
randomly result in an UNKNOWN state with the error message "Session no longer connected".

This was a result of unstable connections to the IPMI boards which should now be handled
in a more stable way.

CMK-1366

Change-Id: I5a9ee67bbaf305295ae68786dccd3ae9fd0bcf86