6918 FIX Availability: Time format is not only affected by time range
In case a availability is displaying a single day time range, the times
of the single entries are formatted without date information.

In case there are annotations or downtimes that affect multiple days the
from and until columns also only showed the start or end time instead of
the date. This was some kind of misleading when trying to understand why
such a downtime matched a given period.

CMK-1353

Change-Id: Iec290d656540add153c5fa28957fcb95c46f6a01