6932 FIX cmk-update-agent: Apply urllib3 bugfix
It has been observed recently that the agent updater might crash
with an error message  like
<pre>IDNAError: The label ... is not a valid A-label</pre>

The cause for this behavior is a bug in the underlying urllib3
python module. However, this has been fixed in a newer urrllib-version
and hence we updated the underlying python modules to newer versions
in order to apply the bugfix to cmk-update-agent.

FEED-3195

Change-Id: I277fdf32962fee68635902239490217482b559cc