6930 Service discovery page can now handle long running tasks
The service discovery page of the configuration GUI could end up
in timeouts when working with devices that need a long time to
process the discovery process. This mostly affected either large
or slow responding SNMP devices.

The discovery page is now initially loading almost immediately
and then starts fetching the content in the background. The
potential long running tasks (full scan or autmatic refresh) are
performed in the background. While these actions are performed,
the GUI shows the already known check table in a read only mode.

Communication with remote pre 1.6 sites is still possible but can
still end up in timeouts.

CMK-1427

Change-Id: I3895aaa5859e1ba893b19b43e6a6e773313b7253