Use open_table() instead of manual begin()/end() pairs.
This is a totally mechanical change, best viewed with whitespace changes
suppressed.  :-) This brings the diff down by an order of magnitude.

Note that there are still a handful of explicit begin()/end() calls: It is a
bit more tricky to get rid of them, and this will be done in a separate
(non-mechanical) change.

Change-Id: Iab730e0449c98122db11ca6c26c67f1c2ffc3285