6929 FIX Distributed WATO login: Protect against some config failures
When building up a distributed Check_MK with distributed WATO, there
are two situations which could end up in problematic situations which
are handled now in a better way:

<ul>
<li>Connecting a central Check_MK Managed Services based site to a CEE
or CRE based site resulted in a broken slave site. When trying to login
a CME site into a CEE site, an error message now prevents the login.</li>
<li>We have added a checkbox to the login dialog to confirm that one
really wants to overwrite the configuration of the remote site with the
central sites configuration.</li>
</ul>

This changes the protocol of the login automation call, but is fully
compatibile with previous versions. We may add additional checks in
the future based on the involved Check_MK editions and versions. The
sites now exchange their versions.

CMK-1370

Change-Id: I2d12dffa4da1fe2c510ac737acbad29127877fec