6934 FIX cmk-update-agent: Remove unnecessary check for python interpreter
Since Check_MK 1.5, the Agent Updater gets deployed as a binary file to
linux systems by default.

Hence it is no longer necessary to check for an available python2 interpreter
on agent installation unless the Agent Updater gets deployed as a script.

This werk adapts the agent's preinstall scripts so that no unnecessary checks are induced anymore.

CMK-1429

Change-Id: I89ca0abb2ef1ded0409725df7caa8249864c1d32