Background jobs: Fixed stop signal handling
The stoppable background jobs could not be stopped in all
situations in a clean way.

The signal processing of the background process was prvented by a
.join() on the execution thread which prevented the python signal
handling from being executed when receiving a SIGTERM to stop the
process.

Cleaned up signal handling in the job process to use the MKTerminate
exception raising up the call stack and make the top run() method
set the final job state.

Cleaned up logging on the way.

Change-Id: I7430a2ad2188dcb00c94437922eeb0866cec664a