7014 Add host name completion in different places
When entering host names in different places, like for example the parent attribute fields of a host
or the nodes fields of a cluster host, a dropdown is shown with the host names matching the already
inserted hostname parts. You may klick on the matching hostname to add it to the input field.

Change-Id: I2ec79aedf073543a5a18363d7536d9e73e10d03c