New host bulk API calls
Added a new API call "add_hosts" which does the same work as the
"add_host" calls with the difference that it may create multiple
hosts.

The request must contain a single key "hosts" which must be a list
containing the single host specifications equal to the add_host call.

Change-Id: I4314b54c7c9868d5ed02fb8b566310a3c0160e04