New API call edit_hosts
Added a new API call "edit_hosts" which does the same work as the
"edit_host" calls with the difference that it may create multiple
hosts.

The request must contain a single key "hosts" which must be a list
containing the single host specifications equal to the edit_host call.

Added tests for the "edit_host" call on the way.

Change-Id: If80aeb83865be64ff94748b3ddf94a1be67f9ed1