Adapted test case for base path redirection.
The previous state was really confusing: We redirect via FOUND (302), but
due to a bug this status code was rendered as "301 Moved Permanently".  The
fix is to keep redirecting via the FOUND status code (which is the same
default as werkzeug/flask) and fix the test case instead.  This is supported
by WebSession.check_redirect(), too.

Anyway, we should have a closer look at this at some point...

Change-Id: I2ab2301953c788aa7fcdb94627b4f761c3dda5f4