6937 FIX Agent Updater: Make executable format rule optional again
Regression since Werk 6934:

With Check_MK 1.5.0p10, baking agents would fail if the rule
"Executable format (Linux)" is not set within the Agent Updater
ruleset. Since this rule is intended to be optional, this werk
restores the normal behavior.

As a workaround for 1.5.0p10, please set the rule to "64bit packaged binary"
to obtain the default setting and make the agents bakeable again.

(Jira: CMK-1429)

Change-Id: I9abf68a95c497a432221fa28ca58b575c3cf1a9f