7029 FIX Skip adding PING service to piggyback only hosts without IP
When a host is configured as "No agent", "No SNMP" and "No IP address"
host it is not useful to add a "PING" service to it in case there are
other services on this host because there is no address to be pinged.

This combination of tags e.g. is used by the Dynamic configuration to
create container hosts or similar.

Change-Id: I0b0ca9c78ef4d027e4129d7d1e929aceff888b9e