6966 Expose HW/SW inventory via HTTP API
Check_MK already has an API to retrieve the HW/SW inventory (see Werk
a new <tt>get_inventory</tt> call has been added to the latter, and this is
now the recommended way to retrieve the HW/SW inventory.

The <tt>get_inventory</tt> call has one mandatory key <tt>hosts</tt>, which
has to be a list of host names for which the inventory should be retrieved.
In addtion, two optional keys are supported:

<ul>
<li><tt>paths</tt>: a list of inventory paths for filtering the inventory tree</li>
<li><tt>site</tt>: the site to use</li>
</ul>

Change-Id: Ibf0efe0bf299021029b084ce076fcb293b7d8d02