6959 Azure Monitoring
A special agent is added to support monitoring of Azure resources.
It can be configured using the rule "Agent Azure Configuration"
in the section "Datasource Programs".
The host which the rule for running the special agent is linked
to will only show overview services.
In particular the service "Azure Agent Info" will show you a list of
currently monitored resource groups. You need to add one Check_MK host
for all theses groups (with the host name being the group name), as
the gathered data for each resource group is distributed via the
piggyback mechanism.

For a list of currently available services you can refer to the check
plungins catalog either in your Check_MK installation or
<a href="cms_check_plugins_catalog.html">online</a>.

CMK-1516

Change-Id: Id3c2e4bb6fbbc4307a10c27141b2d2dbfa60fe9a