Fix update to new predictive levels format
When updating to the new scheme of predictive levels, cached data changed
format and thus old prediction were deleted upon retrieval and new needed
to be calculated on a new step. In the case when the intention is to delete
old predictions on purpose it wasn't checked that they belonged to the old
format but assumed they were already updated. Because during development
one retrieves data always earlier than the maintenance cleanup.

Change-Id: Ie23431019fa8b21fb1a32706314cf8866d654fe0