Fix PS check grab user
The replacement of the usermane of a discovered process in the service
description was neglected. When later doing the replacements of the
match_groups the %u, because is a valid Python replace string, was left
expecting an unsigned int. If there were not match_groups the
replace_service_description would crash.

Change-Id: I5d5522c0cc2983674da34950ef84666360c2712d