7131 FIX mem.win: Fix incorrect representation of absolute memory levels in graphs
Previously, a warn or crit level of less than e.g. 20 GB free space would be displayed
in the graph as a warn or crit level of more than 20 GB used space.
The problem only affected the levels passed to the metrics system for representation,
not the actual levels used for evaluation.

CMK-1543

Change-Id: Iee645048af2d3f85bd19262224fd310c0ae32268