Provide a backport for the check_levels function
Provides the function check_levels_backport in the file
backport.include. It is a backport of the extensively
refactored check_levels function from the master branch.

The aim of the backport is to minimize the differences
of checks between the 1.5.0 and master branch. It should
only be used by features added in the 1.5.0p12 and is
NOT meant as a general purpose include.

The backported version does NOT support predictive levels.

CMK-1518

Change-Id: I2a70fe7e2ba6e2c92c5f8eda2d141d2ac29b2a1e