Simplified cmk.util.paths.
Prepare for future Path() usage on the way.

Note that a few unit tests had to be fixed, they implicitly relied on the
fact that OMD_ROOT is not set and this in turn implies an empty string in
cmk.util.paths.omd_root.  This is a bit fragile and convoluted, perhaps unit
tests should run under a known dummy OMD_ROOT?

Change-Id: I0c7917da9044f3712f05d6edbe87605846782049