6940 FIX Fix OnlyFrom entry in check_mk agent section
Previously, the OnlyFrom entry, that represents the list of
only-from IP entries within the xinetd-configuration,
rendered empty for baked agents.

The cause for this was that the agent was looking for
the service name "check_mk", that only is used by the
raw agent.

The agent is now looking for the right service name set
from the agent bakery, i.e. "check-mk-agent" or a user-specified
name coming from the "name of agent packages" rule.

(Jira: CMK-1540)

Change-Id: I590c7d9f8914923c6e5b3bb12bcab82ed80599f8