7080 FIX Management board: Continue when firmware information can not be fetched
The management board monitoring currently tries to fetch two types of information
from the boards: Sensor states and information about the firmware version. In
case the later could not be fetched the processing of the management board data
was terminated. This has now been changed to continue with the data that could
be fetched.

Change-Id: I26e268e695f87fa867c98aa89659f9f481c433eb