7081 Reworked "Distributed Monitoring" page
The "Distributed Monitoring" pages have been reworked to make it easier
to manage multiple Check_MK site connections.

The list page is now showing less site configuration details. Instead of
these columns status columns have been added to visualize whether or not
your site connections can currently be used.

One column shows the current status of the Livestatus connection, which
is needed by the GUI to gather the monitoring status from the site. In
case your livestatus connection is not configured properly, you may
have a look at the status and hover the icon for more information about
the reason.

One possible reason for a non functional Livestatus connection may be
a TLS connection issue (if you use encrypted Livestatus). You can use
this site to inspect the remote site certificate and establish a trust
with this certificate to allow the GUI to connect with that site.

The replication status column tells you whether or not the configuration
replication connection is working.

CMK-1535

Change-Id: I6939183412f0af033c95f889e1a111bc6be8b546