7057 FIX Allow "coding" comment in notification scripts
Notification scripts can contain the name of the notification
script as a comment after the shebang and the optional encoding
comment. When the encoding comment contained only the word
"coding" but not "encoding" the encoding comment was accidentally
used as the title of the notification script. Now the regular
expression from https://docs.python.org/2.7/reference/lexical_analysis.html
is used to recognize encoding comments. Therefore, "coding" and
"encoding" can now be used.

Change-Id: Ieb2151276a57452d75c3f010bbecaffb037becd1