Disentangled Livestatus Proxy and socket connection settings
The Livestatus socket connection settings (address and port) were
part of the Livestatus Proxy settings in previous versions. This
resulted in several problems.

From the user view: It was not possible to toggle Livestatus Proxy
on and of without having to remember the connection settings to
re-enter it after enabling or disabling the proxy.

From the internal view: There were several places in the code
that had to deal with both kind of socket setting structures in
order to get the destination address.

CMK-1535

Change-Id: If1ec2306b14aff6d5c54822d688c2997ca4cfa30