Increase internal certificate age used for encrypted Livestatus
We decided to set the age of the internally used certificates to
a very long period of 999 years for convenience.

The benefit of a certificate expiration is not really worth the trouble
in our scenario.

In case you have specific security requirements that don't allow such
a long period, you will likely use your own certificates, signed by your
company or some other trusted CA which has a more strict validity period.

CMK-1535

Change-Id: I2838ba706cf98fb4d9fd836d3bb0b5465a74b064