Service discovery: Fix "Updating..." message hiding error result
The "Updating..." message is shown to the user during service discovery
when the request takes more than 2 seconds. In case the discovery
produces a final result this message should be hidden. In case an
unhandled exception occured during discovery within the 2 seconds, the
"Updating..." dialog did overwrite the error message.

Change-Id: Ibf69af2d27f64e9f4852898cbe60b327a38e9d48