DCD: Cleaned up cross component import
The DCD was previously directly executing a remote automation call to a
WATO remote site for gathering information from the remote DCD instance.
At the moment this call is used for fetching phase 1 results from remote
sites.

This was done by using cmk.gui.watolib code, which should not be
imported by another component than the GUI. To solve this we now use a
new Web API call which makes the DCD cleanly operate on the Web API of
the local site. The Web API is then using it's configuration and code to
make the connection to the remote site and returns the result to the
DCD.

Change-Id: I6fe7f76b2b4d66e58a6174e9d7c95005fb9aecae