7084 FIX Add option to delete contact groups with members
When using the LDAP synchronization to populate contact groups in Check_MK,
it was not possible to remove a contact group using the configuration GUI
as long as LDAP users are synchronized to the group.

We have now added a dedicated action to the contact group list which can
be used to remove all members from a contact group. Once you have used
this action, you will be able to remove the contact group from the
configuration.

CMK-1683

Change-Id: I98e896be9c58e5de1cccbfda2266f51da00cf77e