6758 FIX mknotifyd: Fixed crash if the mknotify check sends no data
The check previously crashed, if the agent sends a site name but no data for
this site. If there is no data, this will be displayed now in the service
output. Additionally the service will change it's state to WARN because
there is no data for the last update time of mknotify state file.

Change-Id: Icfab565fb1bcf45218d85413e0470f4df845f899