7088 omd update can now be performed without access to source version
The "omd update" procedure, which is used to perform a version update for a
site, always needed access to the previous version, the new version and the
site.

Since this change it is now possible to perform an update without access
to the old version. This is possible becase we copy the information that
are needed to the Check_MK site during site creation.

This new mechanism can only be used when updating FROM a site that already
implements this werk.

Technical detail:

<ul>
<li>The meta files are saved for the first time during "omd create"</li>
<li>The meta files are updated to the new version during each "omd update"</li>
<li>The files are copied to the sites <tt>.version_meta</tt> directory.</li>
<li>The directory <tt>/omd/versions/[version]/skel is copied.</li>
<li>The file <tt>/omd/versions/[version]/share/omd/skel.permissions is copied.</li>
<li>A file <tt><tt>.version_meta/version</tt> is created.</li>
<li>When an update is performed, the meta files are used if they are available
and up-to-date. In case they don't fit these conditiones, the previous version
files need to be available as before this werk.</li>
<li>

CMK-1603

Change-Id: I4e448e36ce987cbfcedfa47535fab2e9b49b2c19