7164 FIX cmk-update-agent: Add timeout handling for connection to server
Previously, the connection from the agent updater to the Check_MK Server did
not handle timeouts. This could lead to situations where the agent updater
would run forever waiting for a response.

This was particularly noticeable when running the agent updater under Windows,
because the agent updater would fail to start over again, showing an error
like <code>copying plugins\cmk-update-agent.exe to temp\cmk-update-agent.exe failed</code>.

The agent updater now will cancel the connection if there is a connection timeout detected
or if the connection freezes. The timeout limit is set to 120 seconds.<br>
This timeout handling does not limit the overall connection time.

Change-Id: I7d6d49cf099e6c5f326c4dd8dce2c586957facf1