Cleanup module global state cmk.gui.plugins.views.utils.painter_options
The PainterOptions object deals with view (request) specific information
which should never be stored in module scope. As intermediate solution
we store the object in current_app.g["painter_options"].

A better solution would be to create a "view instance" object that holds
the painter options object as helper. This would need a bigger
restructuring of the views code, which we can not focus on at the
moment.

CMK-1626

Change-Id: I6edd86d7c798417c468b0bca5a5b883249ff9667