7174 SEC Apache: Disable TRACE and OPTIONS method
The similar TRACK method is not supported by the site apache at all, so
it does not have to be disabled.

A lot of guides recommend to also disable the OPTIONS method for
production servers.  This HTTP method basically reports which HTTP
Methods that are allowed on the web server. In reality, this is rarely
used for legitimate purposes, but it may grant a potential attacker a
little bit of help and it can be considered a shortcut to find another
hole. For this reason we also disabled the OPTIONS method.

CMK-1639

Change-Id: I57a0b47ba8f7fb2ae0d974ed5383afe56c5418af