7176 FIX Fixed "insecure request warning" message during distributed site remote calls
When executing a remote automation call, for example to rename a host on a remote site
in distributed setups, the message "InsecureRequestWarning: Unverified HTTPS request is
being made" could be visible in the background job output which we don't want to display
there. Insecure configurations are made visibile in the analyze configuration results
instead.

Change-Id: I3e3b3405de3ee42807b65f45077dcefb4b56d455